server/requirements.txt
fastapi
uvicorn
python-socketio[asgi]
python-multipart
pydantic
bcrypt
sqlalchemy
sqlite3